pyproject.toml
[tool.poetry]
name = "sklearn-user-guide"
version = "0.1.0"
description = ""
authors = ["Taisuke Minami <tmnm13009@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.12"
scikit-learn = "^1.4.2"
ipykernel = "^6.29.4"
matplotlib = "^3.8.4"
pandas = "^2.2.2"
seaborn = "^0.13.2"
polars = "^0.20.22"
hydra-core = "^1.3.2"
pyarrow = "^16.0.0"
wandb = "^0.16.6"
plotly = "^5.22.0"
torch = "^2.3.0"
tqdm = "^4.66.4"
lightgbm = "^4.3.0"
optuna = "^3.6.1"
flaml = {extras = ["automl"], version = "^2.1.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
